pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gladius"
version = "1.0.0"
description = "Multi-agent Claude CLI pipeline: Plan → Implement → Review"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "prompt_toolkit>=3.0",
]

[project.scripts]
gladius = "scripts.gladius:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]
